A proposal for a helm chart that supports a monolithic deployment. The same chart is used for monolithic, as was existing for the microservices deployment mode.
To do this a new parameter 'deploymentMode' has been
added to the helm chart. All charts are adapted to
deploy services when applicable according to the
deployment model.

A new template is being added that installs the
monolithic mimir container. All nginx endpoints are
updated to direct to the monolithic service if this
deployment mode has been chosen.

The name of the chart has not been updated from
'mimir-distributed' to 'mimir' yet. This allows for
an easier maintenance as long as this is not merged.
Ideally the rename happens when merging to avoid
further confusion.

I attached a monolithic.yaml yaml file that I have
used to do a monolithic setup.

In monolithic mode, publish not ready addresses and
use these for the frontend.

The read-write mode could be added in the future.

The alert manager is not included in the monolithic
deployment. If you want to run it, you can run it
in a seperate pod by enabling it.

The zone aware replication functionality is supported
in the monolithic deployment with a similar configuration
as required for ingester or store gateway.

I have also added the functionality to enable compression
for the monolithic chart. Ideally this would be embedded
into the complete chart instead. A common configuration is
probably the easiest, if you want grpc message compression,
I guess you want it everywhere. The common configuration
setting can then also be used in the monolithic approach.